pyproject.toml
[project]
name = "langsmith-api-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.0",
    "python-dotenv>=1.1.1",
]
